Add namespace for project and comments
