+++ app/models/project.rb
@@ -1,6 +1,7 @@
 class Project < ApplicationRecord
   belongs_to :user
   has_many  :comments, dependent: :destroy
+  
   validates :description, presence: true
   validates :name,  presence: true, length: { maximum: 20 }
 end
