+++ app/models/user.rb
@@ -5,6 +5,6 @@ class User < ApplicationRecord
 
   has_many :projects, dependent: :destroy
   has_many :comments, dependent: :destroy
-
   has_one_attached :profile_image
+  
 end
