+++ app/views/admin/comments/_comment.html.erb
@@ -1,3 +1,3 @@
 <div class="comments">
-	<%= comment.comment %><br>
+  <%= comment.comment %><br>
 </div>
