+++ app/views/admin/comments/_form.html.erb
@@ -1,0 +1,5 @@
+<%= form_for([:admin, @project, @project.comments.build], remote: true) do |f| %>
+  <h1><%= f.label :comment %></h1><br />
+  <%= f.text_area :comment %><br />
+  <%= f.submit %>
+<% end %>
