+++ app/views/admin/comments/create.js.erb
@@ -1,0 +1,1 @@
+$("#comments-section").append("<%= escape_javascript(render 'admin/comments/comment', comment: @comment) %>")
