+++ app/views/admin/projects/index.html.erb
@@ -1,8 +1,3 @@
-<% if current_user %>
-  <%= link_to 'Sign Out', destroy_user_session_path, method: :delete, class: "btn btn-light" %>
-<% else %>
-  <%= link_to 'Sign Up', new_user_registration_path %>
-<% end %>
 <h1 class="text-center">Create Project</h1>
 <div class="container">
   <table class="table table-borderless table-light">
@@ -19,10 +14,10 @@
           <td><%= project.name %></td>
           <td><%= project.description %></td>
           <td><%= project.created_at%></td>
-          <td><%= link_to 'Show', project, class:"btn btn-light" %></td>
+          <td><%= link_to 'Show', admin_project_path(project), class:"btn btn-light" %></td>
           <% if user_signed_in? %>
-          <td> <%= link_to 'Edit', edit_project_path(project), class:"btn btn-light" %></td>
-          <td><%= link_to 'Delete', project, method: :delete, data: { confirm: 'Are you sure?' } , class:"btn btn-danger"%></td>
+          <td> <%= link_to 'Edit', edit_admin_project_path(project), class:"btn btn-light" %></td>
+          <td><%= link_to 'Delete', admin_project_path(project), method: :delete, data: { confirm: 'Are you sure?' } , class:"btn btn-danger"%></td>
           <% end %>
         </tr>
       <% end %>
