+++ app/views/admin/users/index.html.erb
@@ -15,12 +15,12 @@
           <td><%= user.last_name %></td>
           <td><%= user.created_at%></td>
           <td><%= user.email %></td>
-          <td><%= link_to 'View Project', admin_user_path(user) , class:"btn btn-secondary" %></td>
-          <td> <%= link_to 'Edit', edit_admin_user_path(user) %></td>
-          <td><%= link_to 'Delete', admin_user_path(user), method: :delete, data: { confirm: 'Are you sure?' } %></td>
+          <td><%= link_to 'Show', admin_user_path(user) , class:"btn btn-info"%></td>
+          <td> <%= link_to 'Edit', edit_admin_user_path(user), class:"btn btn-secondary" %></td>
+          <td><%= link_to 'Delete', admin_user_path(user), method: :delete, data: { confirm: 'Are you sure?' } , class: "btn btn-danger"%></td>
         </tr>
       <% end %>
-      All users : <%= @user.count%>     
+      All users : <%= @user.count%>
     </tbody>
   </table>
 </div>
