--- app/views/comments/_form.html.erb
@@ -1,5 +1,0 @@
-<%= form_for([@project, @project.comments.build], remote: true) do |f| %>
-	<h1><%= f.label :comment %></h1><br />
-	<%= f.text_area :comment %><br />
-	<%= f.submit %>
-<% end %>
