--- app/views/comments/create.js.erb
@@ -1,1 +1,0 @@
-$("#comments-section").append("<%= escape_javascript(render 'comments/comment', comment: @comment) %>")
