--- app/views/projects/_form.html.erb
@@ -1,21 +1,0 @@
-<div class="container">
-	<%= form_for @project do |p| %>
-		<% if @project.errors.any? %>
-			<% @project.errors.full_messages.each  do |error| %>
-				<h1> <%= error %></h1>
-			<% end %>
-		<% end %>
-		
-		<div class="form-group">
-    	<label for="exampleInputEmail1">Project Title</label>
-			<%= p.text_field :name, class: "form-control", placeholder: "Enter Project title" %>
-    	<small id="emailHelp" class="form-text text-muted">Enter project tile for understand project type.</small>
-  	</div>
-
-  	<div class="form-group">
-    	<label for="exampleInputPassword1">Project Description</label>
-			<%= p.text_area :description, class: "form-control", placeholder: "Enter your Project Description" %>
-  	</div>
-		<%= p.submit  class:"btn btn-primary" %>
-	<% end %>
-</div>
