--- app/views/projects/show.html.erb
@@ -1,34 +1,0 @@
-	<h1 class="text-center">Project <%= @project.name %></h1>
-	<p id="notice"><%= notice %></p>
-	<%= link_to 'Go Back', projects_path, class: "btn btn-secondary mybtn" %>
-	<%= link_to "Create User", new_admin_user_path, class: "btn btn-secondary mybtn" %>
-<div class="container">	
-	<table class="">
-		<tbody>
-			<tr>
-				<td><strong>Prject Name: </strong></td>
-				<td><%= @project.name %></td>
-			</tr>
-			<tr>
-				<td><strong>Create At: </strong></td>
-				<td><%= @project.created_at %></td>
-			</tr>
-			<tr>
-				<td><strong>Project Description : </strong></td>
-				<td><%= @project.description %></td>
-			</tr>
-		</tbody>
-</table>
-	<div id='comments-section'>
-		<% @project.comments.each do |comment| %>
-			<%= render 'comments/comment', comment: comment %>
-			<%= link_to "Delete", project_comment_path(@project, comment), method: :delete, data: { confirm: 'Are you sure?' } %>
-		<% end %>
-	</div>
-	<div class="comment">
-		Total Comments : <%= @project.comments.count %>	
-	</div>
-	<div id="comment">
-		<%= render 'comments/form'%>
-	</div>
-</div>
