+++ config/routes.rb
@@ -1,15 +1,13 @@
 Rails.application.routes.draw do
-
   devise_for :users
-
+ 
   namespace :admin  do 
     resources :users
+    resources :projects do
+      resources :comments
+  end
 end 
 
-  resources :projects do
-    resources :comments
-end
   resources :clients
-
-  root to: 'projects#index' 
+  root to: 'admin/projects#index' 
 end
